factor maintenance stuff that touches both rabbitmq and azure into its own module
