+++ src/data.js
@@ -2,9 +2,7 @@ let assert = require('assert');
 let Entity = require('azure-entities');
 let taskcluster = require('taskcluster-client');
 let slugid = require('slugid');
-let RabbitManager = require('../lib/rabbitmanager');
 let _ = require('lodash');
-let Debug = require('debug');
 
 /**
  * Entity for keeping track of pulse user credentials
@@ -78,126 +76,6 @@ Namespace.prototype.json = function({cfg, includePassword}) {
 
 Namespace.prototype.username = function() {
   return `${this.namespace}-${this.rotationState}`;
-};
-
-let setPulseUser = async function({username, password, namespace, rabbitManager, cfg}) {
-  await rabbitManager.createUser(username, password, cfg.app.userTags);
-
-  await rabbitManager.setUserPermissions(
-    username,
-    cfg.app.virtualhost,
-    cfg.app.userConfigPermission.replace(/{{namespace}}/, namespace),
-    cfg.app.userWritePermission.replace(/{{namespace}}/, namespace),
-    cfg.app.userReadPermission.replace(/{{namespace}}/, namespace),
-  );
-};
-
-/*
- * Attempt to create a new namespace entry and associated Rabbit user.
- * If the requested namespace exists, update it with any user-supplied settnigs,
- * and return it.
- */
-Namespace.claim = async function({cfg, rabbitManager, namespace, contact, expires}) {
-  let newNamespace;
-  let created;
-
-  try {
-    newNamespace = await Entity.create.call(this, {
-      namespace: namespace,
-      username: namespace,
-      password: slugid.v4(),
-      created: new Date(),
-      expires,
-      rotationState:  '1',
-      nextRotation: taskcluster.fromNow(cfg.app.namespaceRotationInterval),
-      contact,
-    });
-
-    created = true;
-  } catch (err) {
-    if (err.code !== 'EntityAlreadyExists') {
-      throw err;
-    }
-
-    created = false;
-
-    // get the existing row
-    newNamespace = await Entity.load.call(this, {namespace: namespace});
-
-    // If this claim contains different information, update it accordingly
-    if (!_.isEqual(
-      {expires: newNamespace.expires, contact: newNamespace.contact},
-      {expires, contact})) {
-      await newNamespace.modify(entity => {
-        entity.expires = expires;
-        entity.contact = contact;
-      });
-
-      newNamespace = await Entity.load.call(this, {namespace: namespace});
-    }
-  }
-
-  if (created) {
-    // set up the first user as active,
-    await setPulseUser({
-      username: `${namespace}-1`,
-      password: newNamespace.password,
-      namespace, cfg, rabbitManager});
-    // ..and the second user as inactive (empty string means no logins allowed)
-    await setPulseUser({
-      username: `${namespace}-2`,
-      password: '',
-      namespace, cfg, rabbitManager});
-  }
-
-  return newNamespace;
-};
-
-Namespace.expire = async function(now) {
-  assert(now instanceof Date, 'now must be given as option');
-  let count = 0;
-  await Entity.scan.call(this, {
-    expires:          Entity.op.lessThan(now),
-  }, {
-    limit:            250, // max number of concurrent delete operations
-    handler:          (ns) => {
-      count++;
-      return ns.remove(true);
-    },
-  });
-  return count;
-};
-
-Namespace.rotate = async function(now, cfg, rabbitManager) {
-  let count = 0;
-  let debug = Debug('rotate');
-
-  await Entity.scan.call(this, {
-    nextRotation:          Entity.op.lessThan(now),
-  }, {
-    limit:            250, // max number of concurrent modify operations
-    handler:          async (ns) => {
-      count++;
-      debug(`rotating ${ns.namespace}`);
-      let password = slugid.v4();
-      let rotationState = ns.rotationState === '1' ? '2' : '1';
-
-      // modify user in rabbitmq
-      await setPulseUser({
-        username: `${ns.namespace}-${rotationState}`,
-        password,
-        namespace: ns.namespace,
-        cfg, rabbitManager});
-
-      // modify ns in table
-      await ns.modify((entity) => {
-        entity.rotationState = rotationState;
-        entity.nextRotation = taskcluster.fromNow(cfg.app.namespaceRotationInterval),
-        entity.password = password;
-      });
-    },
-  });
-  return count;
 };
 
 module.exports = {Namespace};
